x86 filter for all compression modes. Redo buffer memory allocation.

+++ lrzip_private.h
@@ -406,7 +406,7 @@ struct rzip_control {
 	i64 maxram; // the largest chunk of ram to allocate
 	bool x86filter; // flag for x86 filter use. 0 = no, 1 = yes
 	unsigned char lzma_properties[5]; // lzma properties, encoded
-	unsigned dictSize; // lzma Dictionary size - set in overhead computation
+	i64 dictSize; // lzma Dictionary size - set in overhead computation
 	i64 window;
 	unsigned long flags;
 	i64 ramsize;
